(SIMP-291) Modernize the Ciphers, MACs, and Kex
Added explicit cases for FIPS and non-FIPS mode as well as reasonable
default cases for RHEL7 and below.

Also added support for the KexAlgorithms to get those reasonably in line
with the FIPS requirements out of the box.

Had to pull updated augeas lenses directly from the Augeas project to
get this to work properly. This required creating a sub-RPM for license
compatibility.

SIMP-291 #close #comment Modern Updates

Change-Id: I656d271396534509befcd263be2279d535d44463