(SIMP-560) Changed default ssh server ciphers
A configuration parameter was added to simp_def.yaml to set what ciphers should
be configured for use by the ssh server. Only the strongest ciphers are configured
if ssh::server::params::config_strong_ciphers_only is set to true.
All available ciphers are configured if it is set to false.

simp-core was also updated (simp_def.yaml has the deault added.)

The default is false.