(SIMP-724) Fix error in Puppet 4 compilation
Previous to this commit, the module would not compile in Puppet 4 due to
an used conditional in the ssh::client::add_entry defined type.  The
Puppet 4 compiler disallows code that can never be executed, failing
with the following error: "Evaluation Error: Error while evaluating a
Resource Statement, This 'if' statement has no effect. A value was
produced and then forgotten (one or more preceding expressions may have
the wrong form)"

This commit removes the unused conditional.

SIMP-724 #close

Change-Id: I3a6d37d64bbdbfe74f016bbf05cd6f8df2e38922