(SIMP-2897) Revert ssh::server::conf::trusted_nets
The trusted_nets option has been reverted to 'ALL' since the more
common use case is to install on a virtual machine in one of the
popular cloud infrastructures.

If you do this, then the default of 'simp_options::trusted_nets' is
wrong 100% of the time.

SIMP-2897 #comment Reverted trusted_nets setting
SIMP-2904 #close