(SIMP-2940) Unnecessary permission changes
https://simp-project.atlassian.net/browse/SIMP-2940

The module change the following permissions from the RPM default, which triggers on some security scans. Since security is not enhanced by the changes, it is recommended to leave them at RPM default.

*Set /etc/ssh/moduli permissions to ‘0644’ instead of ‘0600’*
https://bugzilla.redhat.com/show_bug.cgi?id=1043661

*Set /var/empty/sshd to '0711' instead of '0744'*
This actually makes more sense since you'd need to traverse the directory to read the files within.