(SIMP-5718) sshd RhostsRSAAuthentication error
By default, do not configure RhostsRSAAuthentication in sshd_config for
versions of openssh that no longer allow that option (openssh 7.4 and
greater).  The user can still configure this option to match the
outdated STIG checks, but, sshd will emit an error message when
it parses its configuraton file.

SIMP-5718 # close